Randomize thumbnail grid on the asfarasicanremember page
Hack around absence of hooks in the lay theme by hooking into the
`get_term_metadata` filter which the theme uses to get the order of
projects.

This solution is inherently brittle as it assumes on 'private' code and
queries of Lay theme and hence may break with updates. However since the
theme doesn't provide any hooks itself, I couldn't find a better way.

To minimize possible side effects limit scope of the filter to just that
particular page.